Working on key
The key was not working properly at first. At first, it didn't follow the player. It just launched the player across the scene. Once I solved this issue, the key didn't face the correct direction. After lots of time trying to figure this issue out, the key can now properly follow the player. It looks like the player is actually holding the key as well.